refactor(node): rename startup module to store_open
The old name read as process or network startup. This module only
opens a fork-choice store from a directory.
