+++ crates/verity-node/src/store_open.rs
